+++ frontend/src/App.css
@@ -4,7 +4,6 @@
 
 
 .App-header {
-  background-color: #282c34;
   min-height: 100vh;
   display: flex;
   flex-direction: column;
