+++ frontend/src/App.js
@@ -1,21 +1,23 @@
 import React, { useState } from 'react';
-import logo from './logo.svg';
+//import logo from './logo.svg';
 import './App.css';
-import { useEffect } from 'react';
+//import { useEffect } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Login from './components/login';
-import Home from './components/home';
+import Login from './components/Login/login';
+import Home from './components/Home';
+// import Signup from './components/Signup'
 
 
 function App() {
-return (
-    <Router>
-      <Routes>
-        <Route path="/" element={<Login />} />
-        <Route path="/home" element={<Home />} />
-      </Routes>
-    </Router>
- );
-}
+    return (
+      <Router>
+        <Routes>
+          <Route path="/" element={<Login />} />
+          <Route path="/home" element={<Home />} />
+        </Routes>
+      </Router>
+    );
+  }
+  
 
 export default App;
